Moved MongoRepository to the repository pacakage
This is to separate the data definitions from the data accessing, merely an organizational change
