+++ src/main/java/dk/via/sep4/cloud/web/service/WebController.java
@@ -1,7 +1,7 @@
 package dk.via.sep4.cloud.web.service;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
-import dk.via.sep4.cloud.data.MongoRepository;
+import dk.via.sep4.cloud.data.repository.MongoRepository;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
